+++ src/application/overwrites/overwriteRequestPathVariables.test.ts
@@ -27,6 +27,32 @@ describe('overwriteRequestPathVariables', () => {
     expect(result.item.request.url.variables).toMatchSnapshot()
   })
 
+  it('overwrite overwrite the request path variable when overwrite is true', async () => {
+    const overwriteValues = [
+      {
+        key: 'id',
+        value: 'foo-bar-baz',
+        overwrite: true
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
+
+  it('overwrite overwrite the request path variable with an empty when overwrite is true', async () => {
+    const overwriteValues = [
+      {
+        key: 'id',
+        value: '',
+        overwrite: true
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
+
   it('should return unaltered if key not found', async () => {
     const overwriteValues = [
       {
@@ -34,11 +60,98 @@ describe('overwriteRequestPathVariables', () => {
         value: 'foo-bar-baz'
       }
     ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toEqual(pmOperation.item.request.url.variables)
+  })
+
+  it('should remove the request path variable, when remove is true', async () => {
+    const overwriteValues = [
+      {
+        key: 'id',
+        remove: true
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url).toMatchSnapshot()
+  })
+
+  it('should not remove any the request path variable, when a absent key', async () => {
+    const overwriteValues = [
+      {
+        key: 'fake-key',
+        remove: true
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url).toMatchSnapshot()
+  })
+
+  it('should not insert the path variable, if key found', async () => {
+    const overwriteValues = [
+      {
+        key: 'id',
+        value: 'foo-bar-baz',
+        insert: true
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
+
+  it('should insert the path variable, if key not found', async () => {
+    const overwriteValues = [
+      {
+        key: 'add-a-path-variable',
+        value: 'foo-bar-baz'
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
+
+  it('should insert the path variable with true insert option, if key not found', async () => {
+    const overwriteValues = [
+      {
+        key: 'add-a-path-variable',
+        value: 'foo-bar-baz',
+        insert: true
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
+
+  it('should skip the path variable with false insert option, if key not found', async () => {
+    const overwriteValues = [
+      {
+        key: 'add-a-path-variable',
+        value: 'foo-bar-baz',
+        insert: false
+      }
+    ]
 
     const pmOperation = await getPostmanMappedOperation()
     const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
 
-    expect(result.item.request.url.variables).toEqual(pmOperation.item.request.url.variables)
+  it('should insert the path variable with description, if key not found', async () => {
+    const overwriteValues = [
+      {
+        key: 'add-a-path-variable',
+        value: 'foo-bar-baz',
+        description: 'Additional path param'
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
   })
 
   it('should overwrite the request path variable with an empty value', async () => {